+++ src/App.vue
@@ -120,6 +120,7 @@ export default {
 		}
 		this.$store.commit("setLocalStorageLoaded", true);
 		this.$store.dispatch("fetchDummyHtml");
+		this.$store.dispatch("fetchLandingPageDummyHtml");
 	},
 };
 </script>