+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.8.0] - 2025-09-10
+### Added
+- landing page generator
+
+### Changed
+- extend externalElements config example
+- change externalElements iframe example
+
 ## [2.7.0] - 2025-06-06
 ### Added
 - actionMenu.element editor config